Unit Tests: properly patch has_admin_rights
A couple of tests were failing because of a bug in how I was patching the has_admin_rights property and the HTTPClient do_call method. I also now skip the smoke test at the end so it doesn't throw users for a loop - anyone who wants that test will need to comment out the @unittest.skip I put there.